make it easier to implement Payload
